I18N Issues
Your plugin slug is equivalent-mobile-redirect, but your Text Domain is emr-redirect. Change your Text Domain so it is equal to your slug and modify the text domain in all your source files. This change is needed because your "Requires at least" is below 4.6. (see https://developer.wordpress.org/plugins/internationalization/how-to-internationalize-your-plugin/#text-domains)
I corrected all incorrect text domain.
Add a string as localizable.
A UI string is the same as the plugin name, so I made it as disambiguation by context.